create interface for future camera implementations
