+++ chunky/src/java/se/llbit/chunky/renderer/scene/camera/CameraPreset.java
@@ -27,35 +27,35 @@ import se.llbit.chunky.world.Icon;
  */
 public enum CameraPreset {
   ISO_NORTH_WEST("Isometric North-West", Icon.isoNW.fxImage(), -Math.PI / 4, -Math.PI / 4) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PARALLEL);
       camera.setShift(0, 0);
     }
   },
   ISO_NORTH_EAST("Isometric North-East", Icon.isoNE.fxImage(), -3 * Math.PI / 4, -Math.PI / 4) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PARALLEL);
       camera.setShift(0, 0);
     }
   },
   ISO_SOUTH_EAST("Isometric South-East", Icon.isoSE.fxImage(), -5 * Math.PI / 4, -Math.PI / 4) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PARALLEL);
       camera.setShift(0, 0);
     }
   },
   ISO_SOUTH_WEST("Isometric South-West", Icon.isoSW.fxImage(), -7 * Math.PI / 4, -Math.PI / 4) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PARALLEL);
       camera.setShift(0, 0);
     }
   },
   SKYBOX_RIGHT("Skybox Right", Icon.skyboxRight.fxImage(), Math.PI, -Math.PI / 2) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PINHOLE);
       camera.setFoV(90);
@@ -63,7 +63,7 @@ public enum CameraPreset {
     }
   },
   SKYBOX_LEFT("Skybox Left", Icon.skyboxLeft.fxImage(), 0, -Math.PI / 2) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PINHOLE);
       camera.setFoV(90);
@@ -71,7 +71,7 @@ public enum CameraPreset {
     }
   },
   SKYBOX_UP("Skybox Up", Icon.skyboxUp.fxImage(), -Math.PI / 2, Math.PI) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PINHOLE);
       camera.setFoV(90);
@@ -79,7 +79,7 @@ public enum CameraPreset {
     }
   },
   SKYBOX_DOWN("Skybox Down", Icon.skyboxDown.fxImage(), -Math.PI / 2, 0) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PINHOLE);
       camera.setFoV(90);
@@ -87,7 +87,7 @@ public enum CameraPreset {
     }
   },
   SKYBOX_FRONT("Skybox Front (North)", Icon.skyboxFront.fxImage(), -Math.PI / 2, -Math.PI / 2) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PINHOLE);
       camera.setFoV(90);
@@ -95,7 +95,7 @@ public enum CameraPreset {
     }
   },
   SKYBOX_BACK("Skybox Back", Icon.skyboxBack.fxImage(), Math.PI / 2, -Math.PI / 2) {
-    @Override public void apply(Camera camera) {
+    @Override public void apply(MutableCamera camera) {
       camera.setView(yaw, pitch, 0);
       camera.setProjectionMode(ProjectionMode.PINHOLE);
       camera.setFoV(90);
@@ -122,7 +122,7 @@ public enum CameraPreset {
   /**
    * Applies a preset to a camera.
    */
-  abstract public void apply(Camera camera);
+  abstract public void apply(MutableCamera camera);
 
   public Image getIcon() {
     return icon;
