+++ chunky/src/java/se/llbit/chunky/renderer/scene/camera/projection/StereographicProjector.java
@@ -18,7 +18,7 @@ package se.llbit.chunky.renderer.scene.camera.projection;
 
 import java.util.Random;
 
-import se.llbit.chunky.renderer.scene.camera.Camera;
+import se.llbit.chunky.renderer.scene.camera.CameraUtils;
 import se.llbit.math.Vector3;
 
 /**
@@ -31,7 +31,7 @@ public class StereographicProjector implements Projector {
   private final double scale;
 
   public StereographicProjector(double fov) {
-    scale = Camera.clampedFovTan(fov);
+    scale = CameraUtils.clampedFovTan(fov);
   }
 
   @Override public void apply(double x, double y, Random random, Vector3 o, Vector3 d) {
